代码优化2

+++ src/main/java/com/jzpz/repository/ModuleRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by weiQiang on 2016/10/8.
  */
-public interface ModuleRepository extends JpaRepository<Module,Long> {
+public interface ModuleRepository extends JpaRepository<Module,Integer> {
 }
